Restructure CI to use reusable steps in ci_steps.yml
The ci_steps.yml declares a common set of steps for Linux, macOS, and
Windows, and a set of variables named to match the various CMake
configuration options.

The ci_workflow.yml invokes these steps with a succession of builds
that provide collections of CMake settings, passed to the reusable
workflow via the "with:" statement. The "with:" statements provide
default variable settings, allowing each build to specify only
non-default settings, for a compact summary of what's unique about
that job.

Some notes:

* By default, the builds now pre-install Imath and libdeflate, so that
  the main OpenEXR build picks them up. A special build now validates
  the behavior of OPENEXR_FORCE_INTERNAL_IMATH and
  OPENEXR_FORCE_INTERNAL_DEFLATE, which were previously untested.

* For each OS, we now build and test these configurations:
  1. Release
  2. Debug
  3. Static
  4. Threading disabled
  5. pkgconfig, docs, examples, tools disabled. This was also previously not
     tested.
  6. Legacy VFX reference platform compiler(s) and/or os(s)

* The "label:" setting forms the description of the job on the GitHub
  Acions page, prepended with the OS and "build" number, which makes
  it easier to cross reference between the entries on the Actions page
  and the workflow file.

* We no longer build with clang on Linux, since we do a clang build on
  macOS.

Signed-off-by: Cary Phillips <cary@ilm.com>