Rename "HT256" to "htj2k"
This renames the new compression to "htj2k". The discussion on this
concluded that the most significant aspect of the code is the "j2k"
part. The "256" part indicates the number of scanlines per chunk, but
should that should be regarded as the implicit default. Should we ever
introduce a j2k codec with a different scanline count, we can include
that non-default number in the name then.

From the htj2k-beta branch, this renames:

* "ht256" as a compression string name to "htj2k"
* HT256_COMPRESSION to HTJ2K_COMPRESSION
* EXR_COMPRESSION_HT256 to EXR_COMPRESSION_HTJ2K

NOTE: existing files written by applications build against the
htj2k-beta branch will not be readible by applications built with the
library going forward!

Signed-off-by: Cary Phillips <cary@ilm.com>