Properly implement header_only for python module's OpenEXR.File
When `OpenEXR.File` opens a file, it now holds onto the
`MultiPartInputFile` object (deleted by the destructor when the `File`
goes out of scope), so that a subsequent `write` operation can call
`copyPixels` on it to transfer the image data from the input to the
new output file.

A new test in `src/wrappers/python/test/test_unittest.py` validates
the behavior with a muli-part file.

Signed-off-by: Cary Phillips <cary@ilm.com>