Allow arbitrary openjph version
This allows CMake to find whatever version of openjph is available,
since OpenEXR has no strict dependency on any specific openjph
version.

It also set the auto-fetched openjph tag to master, so we're testing
against the bleeding edge, to catch and resolve problems as soon as
possible.

The openjph version will be pinned to a specific release in each OpenEXR
release.

Signed-off-by: Cary Phillips <cary@ilm.com>