Remove openjph from OpenEXR.pc
It's not a part of the public interface, so it's not necessary for
application.

Also, since openexr/src/lib/OpenEXRCore/CMakeLists.txt declares
OpenJPH as private:

 target_link_libraries(OpenEXRCore PRIVATE ${EXR_OPENJPH_LIB})

it is *not* necessary in the `target_link_libraries(PUBLIC)`
declaration in cmake/LibraryDefine.cmake, so this change removes it
there.

This also adds a ci test to validate the OpenEXR.pc.

Signed-off-by: Cary Phillips <cary@ilm.com>