Add only_parts option to python module File read/constructor
This allows reading of only selected parts from a multi-part file. The
`only_parts` argument is a list that can contain either part indices
(int) or part names (str).

An extension of #2149: This can specify the specific parts of a
multi-part EXR file to be loaded to avoid excessive IO when only some
parts of the file is needed by the program. This functionality could
be particular useful when the file is very large and only a very small
portion of it is required in memory.

Co-authored-by: dendenxu <zhenx@zju.edu.cn>
Signed-off-by: Cary Phillips <cary@ilm.com>